requirement.txt
dj-rest-auth==2.2.5
Django==3.2.16
django-allauth==0.51.0
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
python-decouple==3.6
pytz==2022.1
drf-yasg==1.21.3
psycopg2==2.9.3
django-rest-swagger==2.2.0
Pillow==9.2.0
django-filter==22.1